+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/upgrade/UpgradeDestinations.java
@@ -9,8 +9,7 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 
 public enum UpgradeDestinations implements UpgradeDestination {
     IMPORTER(createTranslation("block", ContentIds.IMPORTER.getPath())),
-    EXPORTER(createTranslation("block", ContentIds.EXPORTER.getPath())),
-    INTERFACE(createTranslation("block", ContentIds.INTERFACE.getPath()));
+    EXPORTER(createTranslation("block", ContentIds.EXPORTER.getPath()));
 
     private final Component name;